+++ test/session_test.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-#   Copyright 2015 Futur Solo
+#   Copyright 2016 Futur Solo
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -20,9 +20,10 @@ from futurefinity.utils import *
 import futurefinity.web
 import futurefinity.interface.session
 
+import asyncio
+
 import json
 import nose2
-import asyncio
 import aioredis
 import requests
 import unittest