+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+dev
+---
+- Abstract HTTP Protocol by HTTPRequest and HTTPResponse
+- Fixed Cookie Default Value Bug
+- Fixed SessionInterface Bug
+
 
 v0.1.1
 ------
@@ -10,7 +16,6 @@ v0.1.1
 - Added Keep-Alive Unittest
 - Added Redirect Unittest
 
-
 v0.1.0
 ------
 - StaticFileHandler
@@ -19,7 +24,8 @@ v0.1.0
 - Using Template Interface instead of Overriding Function
 - Introduce Session Interface
 - Introduce Secure Cookie Based Session
-- Introduce Redis Based Sessionv
+- Introduce Redis Based Session
+- Etag Support
 
 0.0.1
 ------